articles

+++ app/Http/Controllers/CategoryController.php
@@ -110,7 +110,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        unlink(public_path('imgs/categories/'.$category->image));
+        if ($category->image && file_exists(public_path('imgs/categories/'. $category->image))) {
+            unlink(public_path('imgs/categories/'.$category->image));
+        }
+        
         $category->delete();
 
         return redirect()->route('admin.categories.index')->with('success', 'Категория успешно удалена');
