+++ app/Providers/AppServiceProvider.php
@@ -5,6 +5,7 @@ namespace App\Providers;
 use Illuminate\Support\ServiceProvider;
 use Schema;
 use App\Category;
+use App\Article;
 use App\Product;
 use App\Manufacture;
 use App\Http\Services\WorkWithImage;
@@ -56,8 +57,11 @@ class AppServiceProvider extends ServiceProvider
                 $old_image = Category::select('image')->find($model->id);
                 // dd($old_image->image);
                 if($model->image != $old_image->image) {
-                    $file = new Filesystem;
-                    $file->delete(public_path().'\imgs\categories\\' . $old_image->image);
+
+                    if (file_exists(public_path().'\imgs\categories\\' . $old_image->image)) {                        
+                        $file = new Filesystem;
+                        $file->delete(public_path().'\imgs\categories\\' . $old_image->image);
+                    }
                     $path = public_path().'\imgs\categories\\';
                     $file = $model->image;
                     $img = new WorkWithImage($file, $path);
@@ -66,6 +70,42 @@ class AppServiceProvider extends ServiceProvider
                 
             }
         });
+
+        Article::creating(function(Article $model){
+            // $model->slug = Str::slug(mb_substr($model->article, 0, 60) . "-", "-");
+            // $double = Article::where('slug', $model->slug)->first();
+            // if ($double) {
+            //     $next_id = Article::select('id')->orderby('id', 'desc')->first()['id'];
+            //     $model->slug .= '-' . ++$next_id;
+            // }
+            if($model->image) {
+                // dd($model);
+                $path = public_path().'\imgs\articles\\';
+                if (!file_exists($path)) {
+                    mkdir($path, 0777);
+                }
+                $file = $model->image;
+                $img = new WorkWithImage($file, $path);
+                $model->image = $img->saveImage();
+            }
+        });
+
+        Article::updating(function(Article $model) {
+            if($model->image) {
+                $old_image = Article::select('image')->find($model->id);
+                if($model->image != $old_image->image) {
+                    if (file_exists(public_path().'\imgs\articles\\' . $old_image->image)) {                        
+                        $file = new Filesystem;
+                        $file->delete(public_path().'\imgs\articles\\' . $old_image->image);
+                    }
+                    $path = public_path().'\imgs\articles\\';
+                    $file = $model->image;
+                    $img = new WorkWithImage($file, $path);
+                    $model->image = $img->saveImage();
+                }                
+            }
+        });
+
 
         // Product::creating(function(Product $model){
         //     $model->slug = Str::slug(mb_substr($model->product, 0, 60), "-");
